Remove singeTask attribute from MainWindow
Previously, due to that launchMode, it was not possible to return to
a ChatWindow from the launcher (recent apps). The launchMode attribute
was removed to enable to the user to resume a chat from a launcher by
returning directly into the previously left ChatWindow.

To prevent a MainWindow-ChatWindow-MainWindow activity stack, the service
notification was changed so that the intent fired by that intent replaces
the current activity stack.

That way, it's possible to resume a chat from the launcher (recent apps list)
and a multiple MainWindow activities from the Notification area are
prevented as well.